+++ docs/source/conf.py
@@ -190,5 +190,6 @@ intersphinx_mapping = {
     'http://static.repoze.org/whatdocs/': None,
     'http://static.repoze.org/whodocs/': None,
     'http://code.gustavonarea.net/repoze.who.plugins.sa/': None,
+    'http://code.gustavonarea.net/repoze.who-friendlyform/': None,
     'http://code.gustavonarea.net/repoze.what.plugins.sql/': None,
     }